Migrate some labels endpoints
