+++ server/service/labels.go
@@ -1,0 +1,70 @@
+package service
+
+import (
+	"context"
+
+	"github.com/fleetdm/fleet/v4/server/fleet"
+)
+
+////////////////////////////////////////////////////////////////////////////////
+// Create Label
+////////////////////////////////////////////////////////////////////////////////
+
+type createLabelRequest struct {
+	fleet.LabelPayload
+}
+
+type createLabelResponse struct {
+	Label labelResponse `json:"label"`
+	Err   error         `json:"error,omitempty"`
+}
+
+func (r createLabelResponse) error() error { return r.Err }
+
+func createLabelEndpoint(ctx context.Context, request interface{}, svc fleet.Service) (interface{}, error) {
+	req := request.(*createLabelRequest)
+
+	label, err := svc.NewLabel(ctx, req.LabelPayload)
+	if err != nil {
+		return createLabelResponse{Err: err}, nil
+	}
+
+	labelResp, err := labelResponseForLabel(ctx, svc, label)
+	if err != nil {
+		return createLabelResponse{Err: err}, nil
+	}
+
+	return createLabelResponse{Label: *labelResp}, nil
+}
+
+func (svc *Service) NewLabel(ctx context.Context, p fleet.LabelPayload) (*fleet.Label, error) {
+	if err := svc.authz.Authorize(ctx, &fleet.Label{}, fleet.ActionWrite); err != nil {
+		return nil, err
+	}
+
+	label := &fleet.Label{}
+
+	if p.Name == nil {
+		return nil, fleet.NewInvalidArgumentError("name", "missing required argument")
+	}
+	label.Name = *p.Name
+
+	if p.Query == nil {
+		return nil, fleet.NewInvalidArgumentError("query", "missing required argument")
+	}
+	label.Query = *p.Query
+
+	if p.Platform != nil {
+		label.Platform = *p.Platform
+	}
+
+	if p.Description != nil {
+		label.Description = *p.Description
+	}
+
+	label, err := svc.ds.NewLabel(ctx, label)
+	if err != nil {
+		return nil, err
+	}
+	return label, nil
+}
