+++ server/service/transport_labels.go
@@ -67,14 +67,6 @@ func decodeApplyLabelSpecsRequest(ctx context.Context, r *http.Request) (interfa
 
 }
 
-func decodeCreateLabelRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	var req createLabelRequest
-	if err := json.NewDecoder(r.Body).Decode(&req.payload); err != nil {
-		return nil, err
-	}
-	return req, nil
-}
-
 func decodeModifyLabelRequest(ctx context.Context, r *http.Request) (interface{}, error) {
 	id, err := uintFromRequest(r, "id")
 	if err != nil {
